requirements.txt
aiohttp~=3.11.18
Django~=5.2.1
djangorestframework~=3.15.2
pip~=25.1.1
attrs~=25.3.0
aiogram~=3.20.0.post0
pytz~=2025.2
idna~=3.10
multidict~=6.4.3
propcache~=0.3.1
asgiref~=3.8.1
sqlparse~=0.5.3
yarl~=1.20.0
aiosignal~=1.3.2
frozenlist~=1.6.0
aiohappyeyeballs~=2.6.1
certifi~=2025.4.26
pydantic~=2.11.4
annotated-types~=0.7.0
django-cors-headers~=4.0.0
magic-filter~=1.0.12
python-dotenv==1.0.1